extract static invoke expressions to folders based on spark grouping
